Update macro_calibration.gms
This file includes a modified calibration mechanism compared to the normal calibration in MESSAGE. Specifically, calibration rates are adjusted so that growth rates are calibrated to demand changes while AEEI rates remain constant. This calibration mechanism is particularly useful when scenarios of rapidly increasing or decreasing demand are examined, which can lead to abrupt changes and inconsistent trajectories for the AEEI factor, especially when the prescribed GDP changes in response to demand changes are not known. However, this mechanism has the limitation that it doesn't allow for calibrating different demand changes across sectors, as growth rates are defined at the node level and are not differentiated across sectors.